requirements.txt
alembic==1.13.1
annotated-types==0.7.0
beautifulsoup4==4.12.3
black==24.4.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
croniter==2.0.5
dagster==1.7.9
dagster-pipes==1.7.9
docstring_parser==0.16
filelock==3.14.0
fsspec==2024.6.0
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.62.2
httplib2==0.22.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
structlog==24.2.0
tabulate==0.9.0
time-machine==2.14.1
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
